move BannerTest to new components folder
