+++ src/frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import BannerTest from "./BannerTest";
+import BannerTest from "./components/BannerTest";
 
 export default function App() {
   return <BannerTest />;
